Add test for GroupNorm module.
